ocr-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
easyocr==1.7.1
Pillow==10.3.0
numpy==1.26.4
pydantic==2.7.4
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
